requirements.txt
Django
djangorestframework
psycopg2-binary
pyyaml
celery
redis
django-filter
python-dotenv
PyYAML